+++ back/src/main/java/com/openclassrooms/mddapi/repositories/TopicsRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.mddapi.repositories;
 import com.openclassrooms.mddapi.models.Topic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TopicsRepository extends JpaRepository<Topic, Integer> {
+public interface TopicsRepository extends JpaRepository<Topic, Long> {
 }
